# api/openapi.yaml
openapi: 3.1.0
info:
  title: Remedence API
  version: 0.1.0
  description: >-
    Canonical API for authenticated tenant-scoped findings, remediation,
    independent verification, protected evidence, reporting, and audit history.
servers:
  - url: /api/v1
    description: Relative Remedence API v1 namespace
...
paths:
...
  /verifications:
    post:
      tags: [Verifications]
      operationId: createVerification
      summary: Start a principal-bound independent verification run
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      description: >-
        Hosted runs bind verifier identity to an administrator-approved execution
        profile and atomically enqueue an isolated job. Local development may omit
        profile_id and record a principal-bound manual run.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVerificationRequest"
      responses:
        "201":
          description: Verification run created in Running state.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            Location:
              schema:
                type: string
            ETag:
              $ref: "#/components/headers/EntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationWithChecks"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /verification-profiles:
    get:
      tags: [Verifications]
      operationId: listVerificationProfiles
      summary: List approved isolated execution profiles
      responses:
        "200":
          description: Safe profile metadata without image commands or secrets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationProfile"
        "500":
          $ref: "#/components/responses/InternalError"
  /verification-jobs/{jobId}:
    parameters:
      - $ref: "#/components/parameters/VerificationJobId"
    get:
      tags: [Verifications]
      operationId: getVerificationJob
      summary: Read queued execution, retry, cancellation, and receipt state
      responses:
        "200":
          description: Tenant-scoped verification job and signed receipts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationJob"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /verification-jobs/{jobId}/cancel:
    parameters:
      - $ref: "#/components/parameters/VerificationJobId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Verifications]
      operationId: cancelVerificationJob
      summary: Request cancellation of a queued or running isolated execution
      responses:
        "200":
          description: Cancellation persisted or requested at the worker lease boundary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationJob"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /verification-jobs/{jobId}/retry:
    parameters:
      - $ref: "#/components/parameters/VerificationJobId"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Verifications]
      operationId: retryVerificationJob
      summary: Requeue a dead-letter isolated execution after operator review
      responses:
        "200":
          description: Dead-letter job returned to the durable queue with prior receipts retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationJob"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /verifications/{verificationId}/checks:
    parameters:
      - $ref: "#/components/parameters/VerificationId"
      - $ref: "#/components/parameters/IfMatch"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Verifications]
      operationId: createVerificationCheck
      summary: Record one check result for a running verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVerificationCheckRequest"
      responses:
        "201":
          description: Verification check recorded.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/EntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCheck"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
  /verifications/{verificationId}/complete:
    parameters:
      - $ref: "#/components/parameters/VerificationId"
      - $ref: "#/components/parameters/IfMatch"
      - $ref: "#/components/parameters/IdempotencyKey"
    post:
      tags: [Verifications]
      operationId: completeVerification
      summary: Complete a verification as Failed or atomically Verified fixed
      description: >-
        A Passed completion records and locks evidence, completes the verification,
        transitions the finding to Verified fixed, and appends audit events in one transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteVerificationRequest"
      responses:
        "200":
          description: Verification completion and resulting finding state.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/EntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCompletion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  headers:
    RequestId:
      description: Correlation identifier for this request.
      schema:
        type: string
        minLength: 1
        maxLength: 80
    EntityTag:
      description: Strong entity version tag for optimistic concurrency.
      schema:
        type: string
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: >-
        Optional tenant- and principal-scoped replay key. Reusing the key for
        the same request replays its successful response for 24 hours; changing
        the operation or payload returns 409.
      schema:
        type: string
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{7,199}$"
    IfMatch:
      name: If-Match
      in: header
      description: Required strong ETag returned for the target resource. Omission returns 428.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
    VerificationId:
      name: verificationId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
    VerificationJobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
  responses:
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The authenticated principal is not authorized for this operation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The requested mutation conflicts with persisted workflow state.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PreconditionFailed:
      description: The supplied entity ETag is stale or targets another resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PreconditionRequired:
      description: The mutation requires If-Match with the current entity ETag.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RequestTooLarge:
      description: JSON request body exceeds 256 KiB.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ServiceUnavailable:
      description: A required protected dependency or artifact is unavailable.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalError:
      description: Unexpected server error without stack traces or local paths.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Severity:
      type: string
      enum: [Critical, High, Medium, Low, Info]
    FindingState:
      type: string
      enum:
        - Needs remediation
        - Remediating
        - Awaiting verification
        - Verification failed
        - Verified fixed
...
    VerificationStatus:
      type: string
      enum: [Queued, Running, Passed, Failed, Cancelled]
    VerificationCheckStatus:
      type: string
      enum: [Pending, Passed, Failed, Skipped]
...
    Finding:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - company_id
        - finding_key
        - title
        - description
        - source
        - severity
        - state
        - owner
        - asset_name
        - detected_at
        - sla_due_at
        - created_at
        - updated_at
        - version
      properties:
        id:
          type: string
        organization_id:
          type: string
        company_id:
          type: string
        finding_key:
          type: string
          examples: [SEC-1042]
        title:
          type: string
        description:
          type: string
        source:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        state:
          $ref: "#/components/schemas/FindingState"
        owner:
          type: string
        asset_name:
          type: string
        detected_at:
          type: string
          format: date-time
        sla_due_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
...
    VerificationRun:
      type: object
      additionalProperties: false
      required:
        - id
        - finding_id
        - remediation_id
        - status
        - method
        - worker_name
        - credential_type
        - execution_source
        - source_revision
        - patch_digest
        - scope
        - result_summary
        - started_at
        - completed_at
        - created_at
        - version
      properties:
        id:
          type: string
        finding_id:
          type: string
        remediation_id:
          type: string
        status:
          $ref: "#/components/schemas/VerificationStatus"
        method:
          type: string
        worker_name:
          type: string
        credential_type:
          type: string
          enum: [session, local-process, worker-profile, legacy-assertion]
        execution_source:
          type: string
        source_revision:
          type: string
        patch_digest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        scope:
          type: string
        result_summary:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
    VerificationCheck:
      type: object
      additionalProperties: false
      required:
        [id, verification_id, sequence, name, status, message, created_at]
      properties:
        id:
          type: string
        verification_id:
          type: string
        sequence:
          type: integer
          minimum: 1
        name:
          type: string
        status:
          $ref: "#/components/schemas/VerificationCheckStatus"
        message:
          type: string
        created_at:
          type: string
          format: date-time
    VerificationWithChecks:
      type: object
      additionalProperties: false
      required: [verification, checks]
      properties:
        verification:
          $ref: "#/components/schemas/VerificationRun"
        checks:
          type: array
          items:
            $ref: "#/components/schemas/VerificationCheck"
        job:
          $ref: "#/components/schemas/VerificationJob"
    VerificationProfile:
      type: object
      additionalProperties: false
      required: [id, timeout_seconds, max_attempts, network]
      properties:
        id:
          type: string
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 3600
        max_attempts:
          type: integer
          minimum: 1
          maximum: 10
        network:
          type: string
          enum: [none]
    VerificationExecutionReceipt:
      type: object
      additionalProperties: false
      required:
        - attempt
        - worker_id
        - profile_id
        - image_digest
        - command_digest
        - started_at
        - completed_at
        - exit_code
        - timed_out
        - output_hash
        - signature
      properties:
        attempt:
          type: integer
          minimum: 1
        worker_id:
          type: string
        profile_id:
          type: string
        image_digest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        command_digest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        exit_code:
          type: [integer, "null"]
        timed_out:
          type: boolean
        output_hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
        signature:
          type: string
          pattern: "^[0-9a-f]{64}$"
    VerificationJob:
      type: object
      additionalProperties: false
      required:
        - id
        - verification_id
        - profile_id
        - status
        - attempt
        - max_attempts
        - timeout_seconds
        - available_at
        - cancellation_requested
        - last_error
        - created_at
        - updated_at
        - receipts
      properties:
        id:
          type: string
        verification_id:
          type: string
        profile_id:
          type: string
        status:
          type: string
          enum: [Queued, Running, Succeeded, Failed, Cancelled, Dead letter]
        attempt:
          type: integer
          minimum: 0
        max_attempts:
          type: integer
          minimum: 1
        timeout_seconds:
          type: integer
          minimum: 1
        available_at:
          type: string
          format: date-time
        cancellation_requested:
          type: boolean
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        receipts:
          type: array
          items:
            $ref: "#/components/schemas/VerificationExecutionReceipt"
    EvidenceItem:
      type: object
      additionalProperties: false
      required:
        - id
        - finding_id
        - verification_id
        - kind
        - label
        - source_reference
        - content_hash
        - artifact_id
        - manifest_hash
        - manifest_signature
        - attested_by
        - metadata
        - created_at
        - locked_at
      properties:
        id:
          type: string
        finding_id:
          type: string
        verification_id:
          type: string
        kind:
          type: string
        label:
          type: string
        source_reference:
          type: string
        content_hash:
          type: string
          pattern: "^[a-f0-9]{64}$"
        artifact_id:
          type: [string, "null"]
        manifest_hash:
          type: [string, "null"]
          pattern: "^[a-f0-9]{64}$"
        manifest_signature:
          type: [string, "null"]
          pattern: "^[a-f0-9]{64}$"
        attested_by:
          type: [string, "null"]
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        locked_at:
          type: [string, "null"]
          format: date-time
...
    VerificationCompletion:
      type: object
      additionalProperties: false
      required: [verification, finding, evidence]
      properties:
        verification:
          $ref: "#/components/schemas/VerificationRun"
        finding:
          $ref: "#/components/schemas/Finding"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
...
    CreateVerificationRequest:
      type: object
      additionalProperties: false
      required:
        [
          finding_id,
          remediation_id,
          method,
          scope,
          source_revision,
          patch_digest,
          checks,
        ]
      properties:
        finding_id:
          type: string
          minLength: 1
          maxLength: 100
        remediation_id:
          type: string
          minLength: 1
          maxLength: 100
        method:
          type: string
          minLength: 1
          maxLength: 200
        source_revision:
          type: string
          minLength: 1
          maxLength: 200
        patch_digest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        scope:
          type: string
          minLength: 1
          maxLength: 1000
        checks:
          type: array
          minItems: 1
          maxItems: 25
          items:
            type: string
            minLength: 1
            maxLength: 300
        profile_id:
          type: string
          minLength: 3
          maxLength: 64
          pattern: "^[a-z0-9][a-z0-9-]+$"
    CreateVerificationCheckRequest:
      type: object
      additionalProperties: false
      required: [sequence, name, status, message]
      properties:
        sequence:
          type: integer
          minimum: 1
          maximum: 100
        name:
          type: string
          minLength: 1
          maxLength: 300
        status:
          type: string
          enum: [Passed, Failed, Skipped]
        message:
          type: string
          maxLength: 2000
    CreateEvidenceItem:
      type: object
      additionalProperties: false
      required: [artifact_id, kind, label, source_reference, metadata]
      properties:
        artifact_id:
          type: string
          minLength: 1
          maxLength: 100
        kind:
          type: string
          minLength: 1
          maxLength: 100
        label:
          type: string
          minLength: 1
          maxLength: 300
        source_reference:
          type: string
          minLength: 1
          maxLength: 1000
        metadata:
          type: object
          additionalProperties: true
...
    CompleteVerificationRequest:
      type: object
      additionalProperties: false
      required: [result, summary]
      properties:
        result:
          type: string
          enum: [Passed, Failed]
        summary:
          type: string
          minLength: 1
          maxLength: 2000
        evidence:
          type: array
          maxItems: 25
          items:
            $ref: "#/components/schemas/CreateEvidenceItem"
...
    ProblemFieldError:
      type: object
      additionalProperties: false
      required: [path, message]
      properties:
        path:
          type: string
        message:
          type: string
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        request_id:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ProblemFieldError"